backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
whitenoise>=6.8,<7.0
psycopg[binary]>=3.1,<4.0
boto3>=1.34,<2.0
django-storages[s3]>=1.14,<2.0
gunicorn>=21.2,<24.0
